Minor passthrough refactoring
