+++ frontend/src/assets/styles/primevue-components-overrides.scss
@@ -71,7 +71,7 @@ label {
 }
 
 /* ------------------------------------ primevue button --------------------------------- */
-/* NOTE: Button colors are now managed by PrimeVue v4 preset in frontend/src/theme/primevue-preset.ts */
+/* NOTE: Button colors are now managed by PrimeVue v4 preset in frontend/src/custom/primevue-preset.ts */
 /* Keeping only layout/structure rules; color rules removed to allow preset CSS variables to apply */
 .#{variables.$primevue-prefix}-button {
     @extend %body-compact-01;
