+++ frontend/src/custom/dialog/dialogPassThrough.scss
