+++ frontend/src/custom/paginator/paginatorPassThrough.scss
